refactor: split widgets classes to multiple files
