+++ packages/DatasourceCustomizer/src/Decorators/Actions/WidgetField/AddressAutocompleteField.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace ForestAdminDatasourceCustomizer\Decorators\Action\WidgetField;
+
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Actions\DynamicField;
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Actions\WidgetField\Widget;
+
+class AddressAutocompleteField extends DynamicField
+{
+    use Widget;
+
+    public function __construct($options)
+    {
+        parent::__construct($options['type'], $options['label']);
+        WidgetField::validateArg($options, 'type', ['type' => 'contains', 'value' => ['String']]);
+        $this->widget = 'AddressAutocomplete';
+    }
+}
