+++ packages/DatasourceCustomizer/src/Decorators/Actions/WidgetField/TextAreaField.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Actions\WidgetField;
+
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Actions\DynamicField;
+use ForestAdmin\AgentPHP\DatasourceCustomizer\Decorators\Actions\Types\FieldType;
+use ForestAdminDatasourceCustomizer\Decorators\Action\WidgetField\WidgetField;
+
+class TextAreaField extends DynamicField
+{
+    use Widget;
+
+    private ?int $rows;
+
+    public function __construct($options)
+    {
+        parent::__construct($options['type'], $options['label']);
+        WidgetField::validateArg($options, 'type', ['type' => 'contains', 'value' => [FieldType::STRING]]);
+        $this->widget = 'TextArea';
+        $this->rows = $options['rows'] ?? null;
+    }
+
+    public function getRows(): ?int
+    {
+        return $this->rows;
+    }
+}
